Casi entrega final
Faltan ciertos retoques, documentarlo todo lo interno, explicar pasos complicados y la documentación escrita.
