+++ Mis pokémons.txt
@@ -1,33 +1,33 @@
-1^bulbasaur^a
-2^ivysaur^h
+1^bulbasaur^h
+2^ivysaur^a
 3^venusaur^a
-4^charmander^h
+4^charmander^a
 5^charmeleon^a
-6^charizard^h
+6^charizard^a
 7^squirtle^h
-8^wartortle^h
-9^blastoise^h
+8^wartortle^a
+9^blastoise^a
 10^caterpie^a
 11^metapod^a
 12^butterfree^h
-13^weedle^h
+13^weedle^a
 14^kakun^h
-15^beedrill^h
+15^beedrill^a
 16^pidgey^a
 17^pidgeotto^h
 18^pidgeot^a
-19^rattat^h
+19^rattat^a
 20^raticate^h
 21^spearow^h
-22^fearow^h
+22^fearow^a
 23^ekans^h
 24^arbok^h
-25^pikachu^a
-26^raichu^a
-27^sandshrew^h
-28^sandslas^a
+25^pikachu^h
+26^raichu^h
+27^sandshrew^a
+28^sandslas^h
 29^nidoran-f^a
-30^nidorin^a
+30^nidorin^h
 31^nidoqueen^a
 32^nidoran-m^a
 33^nidorino^h
@@ -35,119 +35,119 @@
 35^clefairy^h
 36^clefable^a
 37^vulpix^a
-38^ninetales^a
+38^ninetales^h
 39^jigglypuff^h
-40^wigglytuff^h
+40^wigglytuff^a
 41^zubat^a
-42^golbat^h
-43^oddis^a
-44^gloom^h
-45^vileplume^a
-46^paras^h
-47^parasect^a
+42^golbat^a
+43^oddis^h
+44^gloom^a
+45^vileplume^h
+46^paras^a
+47^parasect^h
 48^venonat^h
 49^venomot^a
-50^diglett^a
-51^dugtrio^a
-52^meowt^h
+50^diglett^h
+51^dugtrio^h
+52^meowt^a
 53^persian^h
 54^psyduck^a
-55^golduck^h
-56^mankey^h
-57^primeape^h
-58^growlithe^h
-59^arcanine^a
+55^golduck^a
+56^mankey^a
+57^primeape^a
+58^growlithe^a
+59^arcanine^h
 60^poliwag^a
-61^poliwhirl^a
-62^poliwrat^h
-63^abr^a
+61^poliwhirl^h
+62^poliwrat^a
+63^abr^h
 64^kadabr^a
-65^alakazam^h
-66^machop^a
+65^alakazam^a
+66^machop^h
 67^machoke^a
-68^machamp^h
+68^machamp^a
 69^bellsprout^h
-70^weepinbell^a
-71^victreebel^a
-72^tentacool^a
-73^tentacruel^h
-74^geodude^a
+70^weepinbell^h
+71^victreebel^h
+72^tentacool^h
+73^tentacruel^a
+74^geodude^h
 75^graveler^h
 76^golem^h
-77^ponyt^h
-78^rapidas^a
-79^slowpoke^a
-80^slowbro^h
-81^magnemite^h
-82^magneton^h
-83^farfetchd^a
-84^doduo^h
-85^dodrio^h
-86^seel^a
-87^dewgong^a
+77^ponyt^a
+78^rapidas^h
+79^slowpoke^h
+80^slowbro^a
+81^magnemite^a
+82^magneton^a
+83^farfetchd^h
+84^doduo^a
+85^dodrio^a
+86^seel^h
+87^dewgong^h
 88^grimer^h
 89^muk^h
-90^shellder^h
+90^shellder^a
 91^cloyster^a
 92^gastly^a
-93^haunter^h
-94^gengar^a
+93^haunter^a
+94^gengar^h
 95^onix^a
 96^drowzee^a
 97^hypno^a
-98^krabby^h
+98^krabby^a
 99^kingler^a
-100^voltorb^a
-101^electrode^h
+100^voltorb^h
+101^electrode^a
 102^exeggcute^a
 103^exeggutor^a
-104^cubone^h
-105^marowak^a
-106^hitmonlee^h
-107^hitmonchan^h
-108^lickitung^h
-109^koffing^a
-110^weezing^h
-111^rhyhorn^h
-112^rhydon^a
-113^chansey^h
+104^cubone^a
+105^marowak^h
+106^hitmonlee^a
+107^hitmonchan^a
+108^lickitung^a
+109^koffing^h
+110^weezing^a
+111^rhyhorn^a
+112^rhydon^h
+113^chansey^a
 114^tangel^a
-115^kangaskhan^a
-116^horse^h
-117^seadr^h
+115^kangaskhan^h
+116^horse^a
+117^seadr^a
 118^goldeen^h
-119^seaking^h
+119^seaking^a
 120^staryu^h
 121^starmie^h
-122^mr-mime^a
+122^mr-mime^h
 123^scyther^h
-124^jynx^a
-125^electabuzz^h
-126^magmar^a
-127^pinsir^a
-128^tauros^h
-129^magikarp^h
+124^jynx^h
+125^electabuzz^a
+126^magmar^h
+127^pinsir^h
+128^tauros^a
+129^magikarp^a
 130^gyarados^h
-131^lapras^h
+131^lapras^a
 132^ditto^a
-133^eevee^a
-134^vaporeon^h
+133^eevee^h
+134^vaporeon^a
 135^jolteon^h
-136^flareon^a
-137^porygon^a
-138^omanyte^a
+136^flareon^h
+137^porygon^h
+138^omanyte^h
 139^omastar^h
-140^kabuto^h
+140^kabuto^a
 141^kabutops^h
-142^aerodactyl^h
+142^aerodactyl^a
 143^snorlax^h
-144^articuno^a
-145^zapdos^h
-146^moltres^h
+144^articuno^h
+145^zapdos^a
+146^moltres^a
 147^dratini^h
 148^dragonair^a
 149^dragonite^h
-150^mewtwo^a
+150^mewtwo^h
 151^mew^a
 152^chikorit^h
 153^bayleef^h
@@ -155,47 +155,97 @@
 155^cyndaquil^h
 156^quilav^a
 157^typhlosion^h
-158^totodile^h
+158^totodile^a
 159^croconaw^h
 160^feraligatr^h
-161^sentret^h
-162^furret^a
+161^sentret^a
+162^furret^h
 163^hoothoot^a
-164^noctowl^a
-165^ledyb^h
-166^ledian^h
+164^noctowl^h
+165^ledyb^a
+166^ledian^a
 167^spinarak^a
 168^ariados^a
-169^crobat^a
+169^crobat^h
 170^chinchou^h
 171^lanturn^a
 172^pichu^a
-173^cleff^a
+173^cleff^h
 174^igglybuff^h
-175^togepi^h
-176^togetic^h
-177^natu^h
+175^togepi^a
+176^togetic^a
+177^natu^a
 178^xatu^a
 179^mareep^h
 180^flaaffy^h
-181^ampharos^h
-182^bellossom^h
-183^marill^a
-184^azumarill^a
-185^sudowoodo^h
-186^politoed^h
+181^ampharos^a
+182^bellossom^a
+183^marill^h
+184^azumarill^h
+185^sudowoodo^a
+186^politoed^a
 187^hoppip^h
 188^skiploom^h
-189^jumpluff^a
+189^jumpluff^h
 190^aipom^a
-191^sunkern^h
+191^sunkern^a
 192^sunflor^a
-193^yanm^a
+193^yanm^h
 194^wooper^h
-195^quagsire^h
-196^espeon^h
-197^umbreon^h
+195^quagsire^a
+196^espeon^a
+197^umbreon^a
 198^murkrow^a
-199^slowking^h
-200^misdreavus^h
-569^garbodor^a344^claydol^a
+199^slowking^a
+200^misdreavus^a
+201^unown^a
+202^wobbuffet^h
+203^girafarig^a
+204^pineco^h
+205^forretress^h
+206^dunsparce^a
+207^gligar^a
+208^steelix^a
+209^snubbull^h
+210^granbull^a
+211^qwilfis^h
+212^scizor^a
+213^shuckle^a
+214^heracross^a
+215^sneasel^a
+216^teddiurs^a
+217^ursaring^a
+218^slugm^h
+219^magcargo^a
+220^swinub^a
+221^piloswine^h
+222^corsol^a
+223^remoraid^h
+224^octillery^a
+225^delibird^h
+226^mantine^a
+227^skarmory^a
+228^houndour^h
+229^houndoom^h
+230^kingdr^a
+231^phanpy^h
+232^donphan^h
+233^porygon2^a
+234^stantler^a
+235^smeargle^h
+236^tyrogue^h
+237^hitmontop^h
+238^smoochum^a
+239^elekid^a
+240^magby^h
+241^miltank^a
+242^blissey^h
+243^raikou^a
+244^entei^a
+245^suicune^a
+246^larvitar^a
+247^pupitar^a
+248^tyranitar^a
+249^lugi^h
+250^ho-o^a
+1002^chien-pao^a
