+++ huyeron1.html
@@ -1,0 +1,108 @@
+<html lang="es">
+<head><title>Pokémons Huyeron</title><meta charset="utf-8"></head>
+<body>
+<table border="1">
+<tr><th>#</th><th>ID</th><th>Nombre</th><th>Total Estadísticas</th></tr>
+<tr><td>1</td><td>150</td><td>mewtwo</td><td>680</td></tr>
+<tr><td>2</td><td>250</td><td>ho-o</td><td>680</td></tr>
+<tr><td>3</td><td>248</td><td>tyranitar</td><td>600</td></tr>
+<tr><td>4</td><td>144</td><td>articuno</td><td>580</td></tr>
+<tr><td>5</td><td>145</td><td>zapdos</td><td>580</td></tr>
+<tr><td>6</td><td>146</td><td>moltres</td><td>580</td></tr>
+<tr><td>7</td><td>243</td><td>raikou</td><td>580</td></tr>
+<tr><td>8</td><td>244</td><td>entei</td><td>580</td></tr>
+<tr><td>9</td><td>245</td><td>suicune</td><td>580</td></tr>
+<tr><td>10</td><td>130</td><td>gyarados</td><td>540</td></tr>
+<tr><td>11</td><td>143</td><td>snorlax</td><td>540</td></tr>
+<tr><td>12</td><td>242</td><td>blissey</td><td>540</td></tr>
+<tr><td>13</td><td>6</td><td>charizard</td><td>534</td></tr>
+<tr><td>14</td><td>157</td><td>typhlosion</td><td>534</td></tr>
+<tr><td>15</td><td>9</td><td>blastoise</td><td>530</td></tr>
+<tr><td>16</td><td>103</td><td>exeggutor</td><td>530</td></tr>
+<tr><td>17</td><td>160</td><td>feraligatr</td><td>530</td></tr>
+<tr><td>18</td><td>3</td><td>venusaur</td><td>525</td></tr>
+<tr><td>19</td><td>134</td><td>vaporeon</td><td>525</td></tr>
+<tr><td>20</td><td>196</td><td>espeon</td><td>525</td></tr>
+<tr><td>21</td><td>142</td><td>aerodactyl</td><td>515</td></tr>
+<tr><td>22</td><td>62</td><td>poliwrat</td><td>510</td></tr>
+<tr><td>23</td><td>181</td><td>ampharos</td><td>510</td></tr>
+<tr><td>24</td><td>208</td><td>steelix</td><td>510</td></tr>
+<tr><td>25</td><td>31</td><td>nidoqueen</td><td>505</td></tr>
+<tr><td>26</td><td>34</td><td>nidoking</td><td>505</td></tr>
+<tr><td>27</td><td>68</td><td>machamp</td><td>505</td></tr>
+<tr><td>28</td><td>213</td><td>shuckle</td><td>505</td></tr>
+<tr><td>29</td><td>78</td><td>rapidas</td><td>500</td></tr>
+<tr><td>30</td><td>89</td><td>muk</td><td>500</td></tr>
+<tr><td>31</td><td>212</td><td>scizor</td><td>500</td></tr>
+<tr><td>32</td><td>214</td><td>heracross</td><td>500</td></tr>
+<tr><td>33</td><td>217</td><td>ursaring</td><td>500</td></tr>
+<tr><td>34</td><td>232</td><td>donphan</td><td>500</td></tr>
+<tr><td>35</td><td>139</td><td>omastar</td><td>495</td></tr>
+<tr><td>36</td><td>71</td><td>victreebel</td><td>490</td></tr>
+<tr><td>37</td><td>80</td><td>slowbro</td><td>490</td></tr>
+<tr><td>38</td><td>101</td><td>electrode</td><td>490</td></tr>
+<tr><td>39</td><td>125</td><td>electabuzz</td><td>490</td></tr>
+<tr><td>40</td><td>199</td><td>slowking</td><td>490</td></tr>
+<tr><td>41</td><td>241</td><td>miltank</td><td>490</td></tr>
+<tr><td>42</td><td>26</td><td>raichu</td><td>485</td></tr>
+<tr><td>43</td><td>226</td><td>mantine</td><td>485</td></tr>
+<tr><td>44</td><td>36</td><td>clefable</td><td>483</td></tr>
+<tr><td>45</td><td>97</td><td>hypno</td><td>483</td></tr>
+<tr><td>46</td><td>87</td><td>dewgong</td><td>475</td></tr>
+<tr><td>47</td><td>85</td><td>dodrio</td><td>470</td></tr>
+<tr><td>48</td><td>178</td><td>xatu</td><td>470</td></tr>
+<tr><td>49</td><td>205</td><td>forretress</td><td>465</td></tr>
+<tr><td>50</td><td>227</td><td>skarmory</td><td>465</td></tr>
+<tr><td>51</td><td>234</td><td>stantler</td><td>465</td></tr>
+<tr><td>52</td><td>171</td><td>lanturn</td><td>460</td></tr>
+<tr><td>53</td><td>42</td><td>golbat</td><td>455</td></tr>
+<tr><td>54</td><td>57</td><td>primeape</td><td>455</td></tr>
+<tr><td>55</td><td>106</td><td>hitmonlee</td><td>455</td></tr>
+<tr><td>56</td><td>203</td><td>girafarig</td><td>455</td></tr>
+<tr><td>57</td><td>28</td><td>sandslas</td><td>450</td></tr>
+<tr><td>58</td><td>113</td><td>chansey</td><td>450</td></tr>
+<tr><td>59</td><td>119</td><td>seaking</td><td>450</td></tr>
+<tr><td>60</td><td>221</td><td>piloswine</td><td>450</td></tr>
+<tr><td>61</td><td>22</td><td>fearow</td><td>442</td></tr>
+<tr><td>62</td><td>117</td><td>seadr</td><td>440</td></tr>
+<tr><td>63</td><td>211</td><td>qwilfis</td><td>440</td></tr>
+<tr><td>64</td><td>40</td><td>wigglytuff</td><td>435</td></tr>
+<tr><td>65</td><td>114</td><td>tangel</td><td>435</td></tr>
+<tr><td>66</td><td>207</td><td>gligar</td><td>430</td></tr>
+<tr><td>67</td><td>51</td><td>dugtrio</td><td>425</td></tr>
+<tr><td>68</td><td>105</td><td>marowak</td><td>425</td></tr>
+<tr><td>69</td><td>192</td><td>sunflor</td><td>425</td></tr>
+<tr><td>70</td><td>148</td><td>dragonair</td><td>420</td></tr>
+<tr><td>71</td><td>184</td><td>azumarill</td><td>420</td></tr>
+<tr><td>72</td><td>20</td><td>raticate</td><td>413</td></tr>
+<tr><td>73</td><td>185</td><td>sudowoodo</td><td>410</td></tr>
+<tr><td>74</td><td>247</td><td>pupitar</td><td>410</td></tr>
+<tr><td>75</td><td>2</td><td>ivysaur</td><td>405</td></tr>
+<tr><td>76</td><td>5</td><td>charmeleon</td><td>405</td></tr>
+<tr><td>77</td><td>8</td><td>wartortle</td><td>405</td></tr>
+<tr><td>78</td><td>47</td><td>parasect</td><td>405</td></tr>
+<tr><td>79</td><td>93</td><td>haunter</td><td>405</td></tr>
+<tr><td>80</td><td>156</td><td>quilav</td><td>405</td></tr>
+<tr><td>81</td><td>198</td><td>murkrow</td><td>405</td></tr>
+<tr><td>82</td><td>64</td><td>kadabr</td><td>400</td></tr>
+<tr><td>83</td><td>168</td><td>ariados</td><td>400</td></tr>
+<tr><td>84</td><td>12</td><td>butterfree</td><td>395</td></tr>
+<tr><td>85</td><td>15</td><td>beedrill</td><td>395</td></tr>
+<tr><td>86</td><td>44</td><td>gloom</td><td>395</td></tr>
+<tr><td>87</td><td>137</td><td>porygon</td><td>395</td></tr>
+<tr><td>88</td><td>166</td><td>ledian</td><td>390</td></tr>
+<tr><td>89</td><td>193</td><td>yanm</td><td>390</td></tr>
+<tr><td>90</td><td>61</td><td>poliwhirl</td><td>385</td></tr>
+<tr><td>91</td><td>95</td><td>onix</td><td>385</td></tr>
+<tr><td>92</td><td>108</td><td>lickitung</td><td>385</td></tr>
+<tr><td>93</td><td>180</td><td>flaaffy</td><td>365</td></tr>
+<tr><td>94</td><td>58</td><td>growlithe</td><td>350</td></tr>
+<tr><td>95</td><td>111</td><td>rhyhorn</td><td>345</td></tr>
+<tr><td>96</td><td>109</td><td>koffing</td><td>340</td></tr>
+<tr><td>97</td><td>120</td><td>staryu</td><td>340</td></tr>
+<tr><td>98</td><td>188</td><td>skiploom</td><td>340</td></tr>
+<tr><td>99</td><td>201</td><td>unown</td><td>336</td></tr>
+<tr><td>100</td><td>72</td><td>tentacool</td><td>335</td></tr>
+</table>
+</body>
+</html>
