+++ huyeron2.html
@@ -1,0 +1,58 @@
+<html lang="es">
+<head><title>Pokémons Huyeron</title><meta charset="utf-8"></head>
+<body>
+<table border="1">
+<tr><th>#</th><th>ID</th><th>Nombre</th><th>Total Estadísticas</th></tr>
+<tr><td>101</td><td>216</td><td>teddiurs</td><td>330</td></tr>
+<tr><td>102</td><td>228</td><td>houndour</td><td>330</td></tr>
+<tr><td>103</td><td>231</td><td>phanpy</td><td>330</td></tr>
+<tr><td>104</td><td>96</td><td>drowzee</td><td>328</td></tr>
+<tr><td>105</td><td>81</td><td>magnemite</td><td>325</td></tr>
+<tr><td>106</td><td>86</td><td>seel</td><td>325</td></tr>
+<tr><td>107</td><td>88</td><td>grimer</td><td>325</td></tr>
+<tr><td>108</td><td>133</td><td>eevee</td><td>325</td></tr>
+<tr><td>109</td><td>35</td><td>clefairy</td><td>323</td></tr>
+<tr><td>110</td><td>25</td><td>pikachu</td><td>320</td></tr>
+<tr><td>111</td><td>43</td><td>oddis</td><td>320</td></tr>
+<tr><td>112</td><td>54</td><td>psyduck</td><td>320</td></tr>
+<tr><td>113</td><td>104</td><td>cubone</td><td>320</td></tr>
+<tr><td>114</td><td>152</td><td>chikorit</td><td>318</td></tr>
+<tr><td>115</td><td>79</td><td>slowpoke</td><td>315</td></tr>
+<tr><td>116</td><td>7</td><td>squirtle</td><td>314</td></tr>
+<tr><td>117</td><td>158</td><td>totodile</td><td>314</td></tr>
+<tr><td>118</td><td>63</td><td>abr</td><td>310</td></tr>
+<tr><td>119</td><td>84</td><td>doduo</td><td>310</td></tr>
+<tr><td>120</td><td>92</td><td>gastly</td><td>310</td></tr>
+<tr><td>121</td><td>48</td><td>venonat</td><td>305</td></tr>
+<tr><td>122</td><td>66</td><td>machop</td><td>305</td></tr>
+<tr><td>123</td><td>238</td><td>smoochum</td><td>305</td></tr>
+<tr><td>124</td><td>27</td><td>sandshrew</td><td>300</td></tr>
+<tr><td>125</td><td>60</td><td>poliwag</td><td>300</td></tr>
+<tr><td>126</td><td>69</td><td>bellsprout</td><td>300</td></tr>
+<tr><td>127</td><td>74</td><td>geodude</td><td>300</td></tr>
+<tr><td>128</td><td>147</td><td>dratini</td><td>300</td></tr>
+<tr><td>129</td><td>209</td><td>snubbull</td><td>300</td></tr>
+<tr><td>130</td><td>246</td><td>larvitar</td><td>300</td></tr>
+<tr><td>131</td><td>116</td><td>horse</td><td>295</td></tr>
+<tr><td>132</td><td>204</td><td>pineco</td><td>290</td></tr>
+<tr><td>133</td><td>23</td><td>ekans</td><td>288</td></tr>
+<tr><td>134</td><td>132</td><td>ditto</td><td>288</td></tr>
+<tr><td>135</td><td>179</td><td>mareep</td><td>280</td></tr>
+<tr><td>136</td><td>32</td><td>nidoran-m</td><td>273</td></tr>
+<tr><td>137</td><td>50</td><td>diglett</td><td>265</td></tr>
+<tr><td>138</td><td>165</td><td>ledyb</td><td>265</td></tr>
+<tr><td>139</td><td>21</td><td>spearow</td><td>262</td></tr>
+<tr><td>140</td><td>163</td><td>hoothoot</td><td>262</td></tr>
+<tr><td>141</td><td>167</td><td>spinarak</td><td>250</td></tr>
+<tr><td>142</td><td>183</td><td>marill</td><td>250</td></tr>
+<tr><td>143</td><td>220</td><td>swinub</td><td>250</td></tr>
+<tr><td>144</td><td>235</td><td>smeargle</td><td>250</td></tr>
+<tr><td>145</td><td>41</td><td>zubat</td><td>245</td></tr>
+<tr><td>146</td><td>175</td><td>togepi</td><td>245</td></tr>
+<tr><td>147</td><td>161</td><td>sentret</td><td>215</td></tr>
+<tr><td>148</td><td>10</td><td>caterpie</td><td>195</td></tr>
+<tr><td>149</td><td>13</td><td>weedle</td><td>195</td></tr>
+<tr><td>150</td><td>191</td><td>sunkern</td><td>180</td></tr>
+</table>
+</body>
+</html>
